# Conflicts:
#	drivers/can/can_mcan.c
#	drivers/flash/flash_sam0.c
#	drivers/flash/spi_nor.c
#	drivers/spi/spi_context.h
#	drivers/usb/device/usb_dc_stm32.c
#	dts/bindings/rtc/atmel,sam0-rtc.yaml
#	include/zephyr/drivers/gpio.h
#	include/zephyr/drivers/spi.h
#	include/zephyr/drivers/uart.h
#	include/zephyr/sys/util.h
#	kernel/timeout.c
#	lib/os/reboot.c
#	subsys/settings/src/settings_nvs.c
#	subsys/shell/backends/shell_uart.c
#	subsys/storage/flash_map/flash_map_shell.c
#	west.yml

+++ kernel/timeout.c
@@ -74,7 +74,7 @@ static int32_t next_timeout(void)
 	    ((int64_t)(to->dticks - ticks_elapsed) > (int64_t)INT_MAX)) {
 		ret = MAX_WAIT;
 	} else {
-		ret = MAX(0, to->dticks - ticks_elapsed);
+		ret = (int32_t)MAX(0, (to->dticks - ticks_elapsed));
 	}
 
 #ifdef CONFIG_TIMESLICING
